Review Comments Addressed
Linear Regression -
Part 2 checkup exercise set I Q1: conclusion that there is no relationship between crime rate and price is not correct
Part 2 checkup exercise set I Q2 and Q3: story about scatter plots is missing
Part 2 checkup exercise set I Q3: why are the selected variables interesting? is not answered
Part 2 checkup exercise set II Q1: answer to What was the purpose of taking the log? is not correct
Part 2 checkup exercise set II Q2: histogram of the variables picked in the last section is not plotted
Part 3 checkup exercise set 1 Q1: How does the plot look? Do you notice anything interesting or weird in the plot? Comment on what you see - not answered. Also, we don’t need to compare the scatter plot between fitted and actuals to that between RM and price for this question
Part 3 checkup exercise set 2 Q1: How would you change the model to not fit an intercept term? - not answered
Part 3 checkup exercise set 3 Q3: What are the advantages and disadvantages of each based on your exploration? Based on the information provided by both packages, what advantage does `statsmodels` provide? - not answered
Part 3 checkup exercise set 4 Q1: Interpretation of intercept is missing. For interpretation of coefficient, there is a small correction needed: an increase of one pupil unit in the pupil-teacher ratio is associated with a decrease of 2.16K of house values
Part 3 checkup exercise set 4 Q3: What does F statistic tell you? Not answered
Part 3 checkup exercise set 4 Q4: There seems to be a typo in the comment on the relationship between F-statistic and t-statistic. F-statistic is not equal to R-squared
Cell 77: We can also use OLS api from statsmodels package to answer these questions as it automatically shows these metrics/statistics in the model summary
Part 3 checkup exercise set V Q3: Which variables in the model are significant in predicting house price? Write a story that interprets the coefficients. - not answered. You can use OLS api from statsmodels package to get the p-value for each variable and evaluate their significance
Part 5 checkup exercises Q1: Are there any violations of the model assumptions? - not answered
Part 5 checkup exercises Q6: It appears that the cell above cell number 96, which contains the solution for this part has got mistakenly converted to markdown cell. Please change it to code cell and rerun to print the model summary
Logistic Regression:
Checkup exercise set III: Test score after grid search is exactly the same as what we got from the default model, but the comment says that the result has improved. For the questions Do you think that's a problem? Why do we need cross validation and grid search?,  we need to answer in the context of improvement in test accuracy. Overfitting on training datasets would not result in better test accuracy
Checkup exercise set IV: We also need to score the best model we got from grid search on the test data
Naive Bayes:
Exercise set III: We need to plot the percentage of words in the y-axis instead of count of words. Also, we need to choose min_df as the value where the curve starts to climb steeply, not where it starts to plateau. It should be a much smaller number.
Exercise set IV: Q1 and Q2 are not answered
In cell 22, we should use the inverted mask to filter training data as you can see from cell 19 that we have set values in mask to true for test indexes.
We need to invert the mask in cell 24, 33 and 34 as well when filtering training and testing data